requirements.txt
aiogram>=3.7.0,<4.0.0
SQLAlchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
